+++ EssentialFeed/EssentialFeedTests/Feed Cache/LoadFeedImageDataFromCacheUseCaseTests.swift
@@ -74,7 +74,7 @@ class LoadFeedImageDataFromCacheUseCaseTests: XCTestCase {
 	}
 	
 	func test_loadImageDataFromURL_doesNotDeliverResultAfterSUTInstanceHasBeenDeallocated() {
-		let store = StoreSpy()
+		let store = FeedImageDataStoreSpy()
 		var sut: LocalFeedImageDataLoader? = LocalFeedImageDataLoader(store: store)
 		
 		var received = [FeedImageDataLoader.Result]()
@@ -88,8 +88,8 @@ class LoadFeedImageDataFromCacheUseCaseTests: XCTestCase {
 	
 	// MARK: - Helpers
 	
-	private func makeSUT(currentDate: @escaping () -> Date = Date.init, file: StaticString = #file, line: UInt = #line) -> (sut: LocalFeedImageDataLoader, store: StoreSpy) {
-		let store = StoreSpy()
+	private func makeSUT(currentDate: @escaping () -> Date = Date.init, file: StaticString = #file, line: UInt = #line) -> (sut: LocalFeedImageDataLoader, store: FeedImageDataStoreSpy) {
+		let store = FeedImageDataStoreSpy()
 		let sut = LocalFeedImageDataLoader(store: store)
 		trackForMemoryLeaks(store, file: file, line: line)
 		trackForMemoryLeaks(sut, file: file, line: line)
@@ -129,33 +129,6 @@ class LoadFeedImageDataFromCacheUseCaseTests: XCTestCase {
 		
 		action()
 		wait(for: [exp], timeout: 1.0)
-	}
-	
-	private class StoreSpy: FeedImageDataStore {
-		enum Message: Equatable {
-			case insert(data: Data, for: URL)
-			case retrieve(dataFor: URL)
-		}
-
-		private(set) var receivedMessages = [Message]()
-		private var retrievalCompletions = [(FeedImageDataStore.RetrievalResult) -> Void]()
-
-		func insert(_ data: Data, for url: URL, completion: @escaping (FeedImageDataStore.InsertionResult) -> Void) {
-			receivedMessages.append(.insert(data: data, for: url))
-		}
-		
-		func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
-			receivedMessages.append(.retrieve(dataFor: url))
-			retrievalCompletions.append(completion)
-		}
-		
-		func completeRetrieval(with error: Error, at index: Int = 0) {
-			retrievalCompletions[index](.failure(error))
-		}
-		
-		func completeRetrieval(with data: Data?, at index: Int = 0) {
-			retrievalCompletions[index](.success(data))
-		}
 	}
 	
 }
